cloud_function_trigger/requirements.txt
google-cloud-videointelligence
google-cloud-speech
google-cloud-storage
librosa
jiwer
protobuf